wip. Toggle buttons option.
