+++ core/src/js/components/main-nav/Nav.js
@@ -2,6 +2,7 @@ import { isEsc, isSpace, isEnter } from "../../utilities/keyboard";
 import { createEvent } from '../../utilities/events';
 import NavItem from './NavItem';
 import SubNavItem from './SubNavItem';
+import NavAbstract from './NavAbstract';
 
 /**
  * Represent a navigation menu. May be the top nav or a subnav.
@@ -21,7 +22,7 @@ import SubNavItem from './SubNavItem';
  * @prop {Array}               items      - Instances of NavItem that model
  *                                          each element in the nav
  */
-export default class Nav {
+export default class Nav extends NavAbstract {
 
   /**
    * Create a Nav
@@ -31,27 +32,8 @@ export default class Nav {
    *                                     (NavItem).
    */
   constructor(elem, options) {
-    // Save the passed in configuration options.
-    this.options = options;
-    // Prefixing the random ids.
-    this.idPrefix = options.idPrefix || 'su-';
-    // The nav element.
-    this.elem = elem;
-    // The toggle menu button or none.
-    this.toggle = options.toggle || false;
-    // Set the z-index if configured.
-    if (this.options.zIndex > 1) {
-      this.elem.style.zIndex = this.options.zIndex;
-    }
-    // Remove the class that formats the nav for browsers with javascript disabled.
-    this.elem.classList.remove('no-js');
-    // Give this instance a unique ID.
-    let id = Math.random().toString(36).substr(2, 9);
-    this.id = this.idPrefix + id;
-    this.elem.id = this.idPrefix + id;
-    // Initialize items.
-    this.navItems = [];
-    this.subNavItems = [];
+    super(elem, options);
+    // Create navItems.
     this.createNavItems();
     // Initialize the event listeners.
     this.createEventListeners();
@@ -60,28 +42,7 @@ export default class Nav {
       this.toggle.setNav(this);
     }
     // Add an active class to the children.
-    this.itemActiveClass = options.itemActiveClass || "active";
     this.setActivePath();
-  }
-
-  /**
-   * Create the children nav items.
-   * @return {[type]} [description]
-   */
-  createNavItems() {
-    let items = this.elem.querySelectorAll("#" + this.id + ' > ul > li');
-    items.forEach(
-      item => {
-        // Subnav items have special behaviour.
-        if (item.querySelector(item.tagName + " > ul")) {
-          this.subNavItems.push(new SubNavItem(item, this, this.options));
-        }
-        // NavItems have specific event handling.
-        else {
-          this.navItems.push(new NavItem(item, this, this.options));
-        }
-      }
-    );
   }
 
   /**
@@ -100,39 +61,7 @@ export default class Nav {
   // Event Handlers.
   // -------------------------------------------------------------------------
 
-  /**
-   * Handler for all events attached to an instance of this class. This method
-   * must exist when events are bound to an instance of a class
-   * (vs a function). This method is called for all events bound to an
-   * instance. It inspects the instance (this) for an appropriate handler
-   * based on the event type. If found, it dispatches the event to the
-   * appropriate handler.
-   *
-   * @param {KeyboardEvent} event - The keyboard event.
-   *
-   * @return {*}
-   *   Whatever the dispatched handler returns (in our case nothing)
-   */
-  handleEvent(event) {
-    event = event || window.event;
 
-    // If this class has an onEvent method (onClick, onKeydown) invoke it.
-    const handler = 'on'
-      + event.type.charAt(0).toUpperCase()
-      + event.type.slice(1);
-
-    // What was clicked.
-    const target = event.target || event.srcElement;
-
-    // If the caller passed in their own event handling use that instead.
-    if (this.options.itemEvents && this.options.itemEvents[handler]) {
-      new this.options.itemEvents[handler](event, this);
-    }
-    // Otherwise, check to see if we have an event available.
-    else if (typeof this[handler] === 'function') {
-      return this[handler](event, target);
-    }
-  }
 
   /**
    * [preOpenSubnav description]
@@ -151,18 +80,6 @@ export default class Nav {
   }
 
   /**
-   * Gotta close em all.
-   * @return {[type]} [description]
-   */
-  closeAllSubNavs() {
-    this.subNavItems.forEach(
-      (item, event) => {
-        item.closeSubNav();
-      }
-    );
-  }
-
-  /**
    * [onKeydown description]
    * @param  {[type]} event  [description]
    * @param  {[type]} target [description]
@@ -173,39 +90,6 @@ export default class Nav {
 
     if (isEsc(theKey)) {
       this.closeAllSubNavs();
-    }
-  }
-
-  /**
-   * [setActivePath description]
-   */
-  setActivePath() {
-    if (this.options.activePath !== true) {
-      return;
-    }
-
-    var pathname = window.location.pathname;
-    var anchor = window.location.hash;
-    if (pathname.length) {
-      let currentLink;
-
-      if (!anchor) {
-        currentLink = this.elem.querySelector("a[href*='" + pathname + "']");
-      } else {
-        currentLink = this.elem.querySelector("a[href*='" + anchor + "']");
-      }
-
-      if (currentLink) {
-        while(currentLink) {
-          if (currentLink.getAttribute('id') == this.id) {
-            currentLink = false;
-          }
-          else if (currentLink.tagName == "LI") {
-            currentLink.classList.add(this.itemActiveClass);
-          }
-          currentLink = currentLink.parentNode;
-        }
-      }
     }
   }
