+++ core/src/js/components/main-nav/NavAbstract.js
@@ -1,0 +1,156 @@
+import { isEsc, isSpace, isEnter } from "../../utilities/keyboard";
+import { createEvent } from '../../utilities/events';
+import NavItem from './NavItem';
+import SubNavItem from './SubNavItem';
+
+/**
+ * Represent a navigation menu. May be the top nav or a subnav.
+ *
+ * @prop {HTMLElement|NavItem} elem       - The element that is the nav. May
+ *                                          be a main nav (<nav>) or a subnav
+ *                                          (NavItem).
+ * @prop {Nav}                 topNav     - The instance of Nav that models
+ *                                          the top nav. If this is the top
+ *                                          nav, topNav === this.
+ * @prop {HTMLButtonElement}   toggle     - The <button> in the DOM that
+ *                                          toggles the menu on mobile. NULL
+ *                                          if this is a subnav.
+ * @prop {String}              toggleText - The initial text of the mobile
+ *                                          toggle (so we can reset it when
+ *                                          the mobile nav is closed).
+ * @prop {Array}               items      - Instances of NavItem that model
+ *                                          each element in the nav
+ */
+export default class NavAbstract {
+
+  /**
+   * Create a Nav
+   *
+   * @param {HTMLElement|NavItem} elem - The element that is the nav menu.
+   *                                     May be a main nav (<nav>) or a subnav
+   *                                     (NavItem).
+   */
+  constructor(elem, options) {
+    // Save the passed in configuration options.
+    this.options = options;
+    // Prefixing the random ids.
+    this.idPrefix = options.idPrefix || 'su-';
+    // The nav element.
+    this.elem = elem;
+    // The toggle menu button or none.
+    this.toggle = options.toggle || false;
+    // Set the z-index if configured.
+    if (this.options.zIndex > 1) {
+      this.elem.style.zIndex = this.options.zIndex;
+    }
+    // Give this instance a unique ID.
+    let id = Math.random().toString(36).substr(2, 9);
+    this.id = this.idPrefix + id;
+    this.elem.id = this.idPrefix + id;
+    // Initialize items.
+    this.navItems = [];
+    this.subNavItems = [];
+    // Add an active class to the children.
+    this.itemActiveClass = options.itemActiveClass || "active";
+  }
+
+  /**
+   * Create the children nav items.
+   * @return {[type]} [description]
+   */
+  createNavItems() {
+    let items = this.elem.querySelectorAll("#" + this.id + ' > ul > li');
+    items.forEach(
+      item => {
+        // Subnav items have special behaviour.
+        if (item.querySelector(item.tagName + " > ul")) {
+          this.subNavItems.push(new SubNavItem(item, this, this.options));
+        }
+        // NavItems have specific event handling.
+        else {
+          this.navItems.push(new NavItem(item, this, this.options));
+        }
+      }
+    );
+  }
+
+  /**
+   * Handler for all events attached to an instance of this class. This method
+   * must exist when events are bound to an instance of a class
+   * (vs a function). This method is called for all events bound to an
+   * instance. It inspects the instance (this) for an appropriate handler
+   * based on the event type. If found, it dispatches the event to the
+   * appropriate handler.
+   *
+   * @param {KeyboardEvent} event - The keyboard event.
+   *
+   * @return {*}
+   *   Whatever the dispatched handler returns (in our case nothing)
+   */
+  handleEvent(event) {
+    event = event || window.event;
+
+    // If this class has an onEvent method (onClick, onKeydown) invoke it.
+    const handler = 'on'
+      + event.type.charAt(0).toUpperCase()
+      + event.type.slice(1);
+
+    // What was clicked.
+    const target = event.target || event.srcElement;
+
+    // If the caller passed in their own event handling use that instead.
+    if (this.options.itemEvents && this.options.itemEvents[handler]) {
+      new this.options.itemEvents[handler](event, this);
+    }
+    // Otherwise, check to see if we have an event available.
+    else if (typeof this[handler] === 'function') {
+      return this[handler](event, target);
+    }
+  }
+
+  /**
+   * [setActivePath description]
+   */
+  setActivePath() {
+    if (this.options.activePath !== true) {
+      return;
+    }
+
+    var pathname = window.location.pathname;
+    var anchor = window.location.hash;
+    if (pathname.length) {
+      let currentLink;
+
+      if (!anchor) {
+        currentLink = this.elem.querySelector("a[href*='" + pathname + "']");
+      } else {
+        currentLink = this.elem.querySelector("a[href*='" + anchor + "']");
+      }
+
+      if (currentLink) {
+        while(currentLink) {
+          if (currentLink.getAttribute('id') == this.id) {
+            currentLink = false;
+          }
+          else if (currentLink.tagName == "LI") {
+            currentLink.classList.add(this.itemActiveClass);
+          }
+          currentLink = currentLink.parentNode;
+        }
+      }
+    }
+  }
+
+  /**
+   * Gotta close em all.
+   * @return {[type]} [description]
+   */
+  closeAllSubNavs() {
+    this.subNavItems.forEach(
+      (item, event) => {
+        item.closeSubNav();
+      }
+    );
+  }
+
+}
