+++ core/src/js/components/main-nav/NavItem.js
@@ -1,6 +1,7 @@
 import {normalizeKey} from "../../utilities/keyboard";
-import Nav from './Nav';
 import {createEvent} from '../../utilities/events';
+import Nav from './Nav';
+import NavItemAbstract from './NavItemAbstract';
 
 /**
  * Represent an item in a navigation menu. May be a direct link or a subnav
@@ -12,7 +13,7 @@ import {createEvent} from '../../utilities/events';
  * @prop {HTMLLIElement}   link   - the <a> in the DOM that is contained in
  *                                  item (the <li>).
  */
-export default class NavItem {
+export default class NavItem extends NavItemAbstract {
 
   /**
    * Create a NavItem
@@ -21,75 +22,9 @@ export default class NavItem {
    *                                  be a main nav (<nav>) or a subnav (Nav).
    */
   constructor(item, nav, options) {
-    this.options = options;
-    this.item = item;
-    this.nav = nav;
-    this.link = this.item.querySelector(this.item.tagName + ' > a');
+    super(item, nav, options);
     this.link.addEventListener('keydown', this);
     this.link.addEventListener('click', this);
-  }
-
-  // -------------------------------------------------------------------------
-  // Event Handlers.
-  // -------------------------------------------------------------------------
-
-  /**
-   * Handler for all events attached to an instance of this class. This method
-   * must exist when events are bound to an instance of a class
-   * (vs a function). This method is called for all events bound to an
-   * instance. It inspects the instance (this) for an appropriate handler
-   * based on the event type. If found, it dispatches the event to the
-   * appropriate handler.
-   *
-   * @param {KeyboardEvent} event - The keyboard event.
-   *
-   * @return {*}
-   *   Whatever the dispatched handler returns (in our case nothing)
-   */
-  handleEvent(event) {
-    event = event || window.event;
-
-    // If this class has an onEvent method (onClick, onKeydown) invoke it.
-    let handler = 'on'
-      + event.type.charAt(0).toUpperCase()
-      + event.type.slice(1);
-
-    // What was clicked.
-    let target = event.target || event.srcElement;
-
-    // If the caller passed in their own event handling use that instead.
-    if (this.options.itemEvents && this.options.itemEvents[handler]) {
-      new this.options.itemEvents[handler](event, this);
-    }
-    // Otherwise, check to see if we have an event available.
-    else if (typeof this[handler] === 'function') {
-      return this[handler](event, target);
-    }
-  }
-
-  /**
-   * Handler for keydown events. keydown is bound to all NavItem's.
-   * Dispatched from this.handleEvent().
-   *
-   * @param {KeyboardEvent} event - The keyboard event object.
-   * @param {HTMLElement} target  - The HTML element target.
-   */
-  onKeydown(event, target) {
-    let theKey = event.key || event.keyCode;
-    let normalized = normalizeKey(theKey);
-
-    // We don't know about that key.
-    if (!normalized) {
-      return;
-    }
-
-    // Prepare a dynamic handler.
-    let handler = 'onKeydown'
-      + normalized.charAt(0).toUpperCase()
-      + normalized.slice(1);
-    if (typeof this[handler] === 'function') {
-      return this[handler](event, target);
-    }
   }
 
   /**
