+++ core/src/scss/components/secondary-nav/index.scss
@@ -6,4 +6,5 @@
 
 @import
 'secondary-nav',
-'secondary-nav--light';
+'secondary-nav--light',
+'secondary-nav--toggles';
